backend/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-JWT-Extended>=4.4
Flask-Admin>=1.6.0
APScheduler>=3.10
requests>=2.28
python-dotenv>=1.0.0
gunicorn

flask-cors>=3.0
